Compra Itens Request 3 (Correto)

+++ api/src/main/java/algar/desafio/api/usuario/DadosAtualizacaoUsuario.java
@@ -6,12 +6,12 @@ import jakarta.validation.constraints.NotNull;
 
 public record DadosAtualizacaoUsuario(
 
-                @NotNull Long id,
-                String nome,
-                String cpf,
-                String email,
-                double saldo,
-                int itens
+        @NotNull Long id,
+        String nome,
+        String cpf,
+        String email,
+        double saldo,
+        Long itens
 
 ) {
