+++ api/src/main/resources/db/migration/V3__alter-table-usuarios.sql
@@ -1,1 +1,1 @@
-alter table usuarios add itens int;
+alter table usuarios add itens bigint;
